pyproject.toml
[tool.poetry]
name = "django-project"
version = "0.1.0"
description = ""
authors = ["LekaAli <atshoane@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.0"
djangorestframework = "^3.13.1"
coreapi = "^2.3.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
